session 3 demo 3/joke-generator/requirements.txt
langchain
langchain-google-genai
langserve[all]
python-dotenv
uvicorn
fastapi
